+++ CHANGELOG.md
@@ -4,8 +4,12 @@
 
 # Changelog
 
-## 7.0.0
+## 7.0.0 (unreleased)
 * **Features**
+  * add GitHub Actions, GitLab CI and Jenkins environment variable support
+    * GitHub Actions: if `$GITHUB_ACTIONS == true` `$GITHUB_REF` is considered 
+    * GitLab CI: if `GITLAB_CI == true` `$CI_COMMIT_BRANCH` and `$CI_COMMIT_TAG` are considered 
+    * Jenkins: if `JENKINS_HOME` is set `BRANCH_NAME` and `TAG_NAME` are considered 
   * Simplify xml configuration (also see BREAKING CHANGES), e.g.
       ```xml
       <configuration xmlns="https://github.com/qoomon/maven-git-versioning-extension"